# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Kalidou GUISSE",
        "description": "Documentation de mon API Laravel pour la gestion des clients et comptes",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/admins/comptes": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Lister tous les comptes avec pagination et filtres (Admin)",
                "description": "Retourne la liste paginée de tous les comptes avec support pour la pagination, le tri et les filtres.",
                "operationId": "60314e3f98b2d9e5c1fcb64e00ef304b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "created_at",
                                "solde",
                                "numeroCompte"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher dans numeroCompte, nom ou prenom du client",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des comptes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompteResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admins": {
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Créer un nouvel admin",
                "description": "Crée un nouvel admin avec les informations fournies.",
                "operationId": "83528520c2f718a850babe9cded3828a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "281f3aa6-bffe-41bd-8c31-2b02a1e49275"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "23886990999492"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Kilback Laury"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ],
                        "example": "cheque"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1558.59
                    },
                    "devise": {
                        "type": "string",
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T11:37:37.000000Z"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "actif"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": "Inactivité de 30+ jours",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-22T11:37:37.082038Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CompteResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Compte"
                        }
                    },
                    "pagination": {
                        "properties": {
                            "currentPage": {
                                "type": "integer",
                                "example": 1
                            },
                            "totalPages": {
                                "type": "integer",
                                "example": 5
                            },
                            "totalItems": {
                                "type": "integer",
                                "example": 50
                            },
                            "hasNext": {
                                "type": "boolean",
                                "example": true
                            },
                            "hasPrevious": {
                                "type": "boolean",
                                "example": false
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=1"
                            },
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=5"
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=2",
                                "nullable": true
                            },
                            "previous": {
                                "type": "string",
                                "example": null,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides"
                    },
                    "errors": {
                        "properties": {
                            "nom": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Le champ nom est obligatoire."
                                ]
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Le champ email doit être une adresse email valide."
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "code": {
                        "type": "integer",
                        "example": 422
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "Admin": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "c7917097-4734-4dfd-a462-b6c8f8d4aea3"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Kilback"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Laury"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "brigitte.kuhn@example.net"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+1-201-736-0670"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T11:37:37.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T11:37:37.000000Z"
                    }
                },
                "type": "object"
            },
            "AdminResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Admin créé avec succès !"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Admin"
                    }
                },
                "type": "object"
            },
            "AdminCreateRequest": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "telephone"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Kilback"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Laury"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "brigitte.kuhn@example.net"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+1-201-736-0670"
                    }
                },
                "type": "object"
            }
        }
    },
...
}